move /ds to /builder/data-sources
