+++ front/pages/w/[wId]/builder/data-sources/[name]/settings.tsx
